+++ bittytax/conv/parsers/etherscan.py
@@ -5,7 +5,7 @@ from decimal import Decimal
 
 from ..out_record import TransactionOutRecord
 from ..dataparser import DataParser
-from ..exceptions import UnknownAddressError
+from ..exceptions import DataFilenameError
 
 WALLET = "Ethereum"
 
@@ -36,6 +36,23 @@ def parse_etherscan(data_row, _parser, _filename):
                                                  fee_asset="ETH",
                                                  wallet=WALLET)
 
+def parse_etherscan_internal(data_row, _parser, _filename):
+    in_row = data_row.in_row
+    data_row.timestamp = DataParser.parse_timestamp(int(in_row[2]))
+
+    if Decimal(in_row[10]) > 0:
+        data_row.t_record = TransactionOutRecord(TransactionOutRecord.TYPE_DEPOSIT,
+                                                 data_row.timestamp,
+                                                 buy_quantity=in_row[10],
+                                                 buy_asset="ETH",
+                                                 wallet=WALLET)
+    elif Decimal(in_row[11]) > 0:
+        data_row.t_record = TransactionOutRecord(TransactionOutRecord.TYPE_WITHDRAWAL,
+                                                 data_row.timestamp,
+                                                 sell_quantity=in_row[11],
+                                                 sell_asset="ETH",
+                                                 wallet=WALLET)
+
 def parse_etherscan_tokens(data_row, _parser, filename):
     in_row = data_row.in_row
     data_row.timestamp = DataParser.parse_timestamp(int(in_row[1]))
@@ -53,7 +70,7 @@ def parse_etherscan_tokens(data_row, _parser, filename):
                                                  sell_asset=in_row[8],
                                                  wallet=WALLET)
     else:
-        raise UnknownAddressError
+        raise DataFilenameError(filename, "Ethereum address")
 
 def parse_etherscan_nfts(data_row, _parser, filename):
     in_row = data_row.in_row
@@ -72,10 +89,10 @@ def parse_etherscan_nfts(data_row, _parser, filename):
                                                  sell_asset=in_row[8],
                                                  wallet=WALLET)
     else:
-        raise UnknownAddressError
+        raise DataFilenameError(filename, "Ethereum address")
 
 DataParser(DataParser.TYPE_EXPLORER,
-           "Etherscan (Ethereum)",
+           "Etherscan (Ethereum Transactions)",
            ['Txhash', 'Blockno', 'UnixTimestamp', 'DateTime', 'From', 'To', 'ContractAddress',
             'Value_IN(ETH)', 'Value_OUT(ETH)', None, 'TxnFee(ETH)', 'TxnFee(USD)',
             'Historical $Price/Eth', 'Status', 'ErrCode'],
@@ -83,12 +100,20 @@ DataParser(DataParser.TYPE_EXPLORER,
            row_handler=parse_etherscan)
 
 DataParser(DataParser.TYPE_EXPLORER,
-           "Etherscan (Ethereum)",
+           "Etherscan (Ethereum Transactions)",
            ['Txhash', 'Blockno', 'UnixTimestamp', 'DateTime', 'From', 'To', 'ContractAddress',
             'Value_IN(ETH)', 'Value_OUT(ETH)', None, 'TxnFee(ETH)', 'TxnFee(USD)',
             'Historical $Price/Eth', 'Status', 'ErrCode', 'PrivateNote'],
            worksheet_name="Etherscan",
            row_handler=parse_etherscan)
+
+DataParser(DataParser.TYPE_EXPLORER,
+           "Etherscan (Internal Transactions)",
+           ['Txhash', 'Blockno', 'UnixTimestamp', 'DateTime', 'ParentTxFrom', 'ParentTxTo',
+            'ParentTxETH_Value', 'From', 'TxTo', 'ContractAddress', 'Value_IN(ETH)',
+            'Value_OUT(ETH)', None, 'Historical $Price/Eth', 'Status', 'ErrCode', 'Type'],
+           worksheet_name="Etherscan",
+           row_handler=parse_etherscan_internal)
 
 DataParser(DataParser.TYPE_EXPLORER,
            "Etherscan (ERC-20 Tokens)",